destroyObject function for destroying in-game objets.
Trying to restructure things a bit. Containers are just hidden, because they may have things taken in/out of them and need to re-exist. For now. (Cannot handle animated/multipart containers, incidentally.)

Perhaps searching for the destroyed-to-be object should be done inside the function, and there should be another function for direct object destruction.