Move character animations playable via interactions. Move their definition into data from code.
There's still two required default animations, speak and idle, though they're defined in data as well.

The animation frames (pngs themselves) are put in the character layer in images json. The new character_animations.json file defines the name, frames and frame durations of the animation.

There's a new interaction command, play_character_animation, which needs to define the name of the animation and the length of playing it before timing out back to the idle animation. As an example, clicking the exit doors in LZ now display the panic animation, which is a bit silly but will do as an example.

In the code, the speak animation part of setMonologue was split into a new more general playCharacterAnimation function, the monologue_timeout variable and its handling was named to character_animation_timeout, and the stopTalking function is now a stopCharacterAnimations function.

Closes #42. Closes #41. Closes #40. Closes #19. Wow!