Move sequence data to sequences.json and music data to music.json.
Not quite sure if it's useful to have a JSON for music, but it was a non-intrusive change to the current code.

objects.json still has data for end_layer, start_layer and start_layer_menu. Note that end_layer used to have a menu, and that rooms can be set to have music via objects.json. I think this stuff is going to be in images.json.