Fix character's head being under fade layer in room transitions.
Introduced in 5be63c. In do_transition, it was assumed that slow_fade meant a sequence, and no slow_fade meant a room transition and character_layer would be moved to top, so that it wouldn't be under the fade layer.

Stopgap fix now moves the character layer on top after the intro, but this should be handled more robustly once you have the ability to play sequences at any point of the game. A good way to handle this might be to have two fade layers: one for sequences and one for room transitions. You wouldn't need to juggle layer poistions, like now.