Hide menu during sequence playing, instead of do_transition.
Less than ideal, it'd be nice to hide menu when the game is started, but I couldn't get the delay right so it'd happen after the fadeout.

Don't hide text_layer (speech bubble), I don't think it's needed as any speech bubble is cleared with each click on screen (or menu).