Separate room and full screen fade layers.
Room transitions use fade_layer_room and fade_room, sequence etc transitions use fade_layer_full and fade_full.

It is not necessary to move the layers around to draw in the correct order. Everything is and must be ordered in correct drawing order in images.json. Bottom stuff first.

Handling the final image in a sequence (if it transitions, etc) isn't pretty.

Closes #87. Fixes #52.