Don't handle transitions inside play_sequence.
Instead call them after play_sequence using the returned delay. The delay doesn't include the fade-in of the final fade; this is so that the do_transition call can be done between the fade-out and fade-in.

Moving towards cutscenes that can be played via interactions. do_transition command will be needed to call separately now. The adventure designer will likely need to count the necessary timeout themself, unfortunately.