Console warnings if cannot find object from stage.
getObject function, a wrapper of sorts for stage.get. Logs error messages in console if it cannot find the object by the specified id.

Helpful with typos in jsons, and also gives some errors if an object required by the kiigame.js script itself is missing. Not used for optional start_layer, splash_screen and starting menu.