build: configure Netlify build command in the repo (#174)
Because the build command was only in Netlify UI, I missed that changing "npm run dev" to "npm run build-dev" broke the preview deployments. Configure it in the repo so that it's harder to forget about.

Confirmed to work in Netlify UI - deploy output states it found the config file, and everything looks ok.